- Too many changes to mention, see ChangeLog file.
git-svn-id: https://xchat.svn.sourceforge.net/svnroot/xchat@824 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1

+++ src/common/network.c
@@ -22,6 +22,7 @@
 #include <string.h>
 #include <stdio.h>
 #include <unistd.h>
+#include <glib.h>
 
 #include "../../config.h"				  /* grab USE_IPV6 define */
 
@@ -48,7 +49,7 @@ net_set_socket_options (int sok)
 }
 
 char *
-net_ip (unsigned long addr)
+net_ip (guint32 addr)
 {
 	struct in_addr ia;
